+++ src/Contracts/Formatter.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Yajra\DataTables\Contracts;
+
+interface Formatter
+{
+    /**
+     * @param string $value
+     * @param mixed $row
+     * @return string
+     */
+    public function format($value, $row);
+}