+++ src/DataTableAbstract.php
@@ -10,6 +10,7 @@ use Illuminate\Support\Str;
 use Illuminate\Support\Traits\Macroable;
 use Psr\Log\LoggerInterface;
 use Yajra\DataTables\Contracts\DataTable;
+use Yajra\DataTables\Contracts\Formatter;
 use Yajra\DataTables\Exceptions\Exception;
 use Yajra\DataTables\Processors\DataProcessor;
 use Yajra\DataTables\Utilities\Helper;
@@ -146,6 +147,11 @@ abstract class DataTableAbstract implements DataTable, Arrayable, Jsonable
     protected $serializer;
 
     /**
+     * @var array
+     */
+    protected $searchPanes = [];
+
+    /**
      * Can the DataTable engine be created with these parameters.
      *
      * @param mixed $source
@@ -180,6 +186,29 @@ abstract class DataTableAbstract implements DataTable, Arrayable, Jsonable
         $this->extraColumns[] = $name;
 
         $this->columnDef['append'][] = ['name' => $name, 'content' => $content, 'order' => $order];
+
+        return $this;
+    }
+
+    /**
+     * @param string|array $columns
+     * @param mixed|\Yajra\DataTables\Contracts\Formatter $formatter
+     * @return $this
+     * @throws \Exception
+     */
+    public function formatColumn($columns, $formatter)
+    {
+        if (is_string($formatter) && class_exists($formatter)) {
+            $formatter = app($formatter);
+        }
+
+        if (! $formatter instanceof Formatter) {
+            throw new \Exception('$formatter must be an instance of '. Formatter::class);
+        }
+
+        foreach ((array) $columns as $column) {
+            $this->addColumn($column . '_formatted', $formatter);
+        }
 
         return $this;
     }
@@ -657,7 +686,16 @@ abstract class DataTableAbstract implements DataTable, Arrayable, Jsonable
         }
 
         $this->columnSearch();
+        $this->searchPanesSearch();
         $this->filteredRecords = $this->isFilterApplied ? $this->filteredCount() : $this->totalRecords;
+    }
+
+    /**
+     * Perform search using search pane values.
+     */
+    protected function searchPanesSearch()
+    {
+        // Add support for search pane.
     }
 
     /**
@@ -770,6 +808,10 @@ abstract class DataTableAbstract implements DataTable, Arrayable, Jsonable
 
         if ($this->config->isDebugging()) {
             $output = $this->showDebugger($output);
+        }
+
+        foreach ($this->searchPanes as $column => $searchPane) {
+            $output['searchPanes']['options'][$column] = $searchPane['options'];
         }
 
         return new JsonResponse(
@@ -923,5 +965,27 @@ abstract class DataTableAbstract implements DataTable, Arrayable, Jsonable
     protected function getPrimaryKeyName()
     {
         return 'id';
+    }
+
+    /**
+     * Add a search pane options on response.
+     *
+     * @param string $column
+     * @param mixed $options
+     * @param callable|null $builder
+     * @return $this
+     */
+    public function searchPane($column, $options, callable $builder = null)
+    {
+        $options = value($options);
+
+        if ($options instanceof Arrayable) {
+            $options = $options->toArray();
+        }
+
+        $this->searchPanes[$column]['options'] = $options;
+        $this->searchPanes[$column]['builder'] = $builder;
+
+        return $this;
     }
 }